Move resource functions around further
