+++ pkg/provider/cloud/azure/security_group.go
@@ -1,0 +1,201 @@
+/*
+Copyright 2021 The Kubermatic Kubernetes Platform contributors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package azure
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2021-03-01/network"
+	"github.com/Azure/go-autorest/autorest/to"
+
+	kubermaticv1 "k8c.io/kubermatic/v2/pkg/crd/kubermatic/v1"
+)
+
+// ensureSecurityGroup will create or update an Azure security group. The call is idempotent.
+func (a *Azure) ensureSecurityGroup(cloud kubermaticv1.CloudSpec, location string, clusterName string, portRangeLow int, portRangeHigh int, credentials Credentials) error {
+	sgClient, err := getSecurityGroupsClient(cloud, credentials)
+	if err != nil {
+		return err
+	}
+
+	parameters := network.SecurityGroup{
+		Name:     to.StringPtr(cloud.Azure.SecurityGroup),
+		Location: to.StringPtr(location),
+		Tags: map[string]*string{
+			clusterTagKey: to.StringPtr(clusterName),
+		},
+		SecurityGroupPropertiesFormat: &network.SecurityGroupPropertiesFormat{
+			Subnets: &[]network.Subnet{
+				{
+					Name: to.StringPtr(cloud.Azure.SubnetName),
+					ID:   to.StringPtr(assembleSubnetID(cloud)),
+				},
+			},
+			// inbound
+			SecurityRules: &[]network.SecurityRule{
+				{
+					Name: to.StringPtr("ssh_ingress"),
+					SecurityRulePropertiesFormat: &network.SecurityRulePropertiesFormat{
+						Direction:                network.SecurityRuleDirectionInbound,
+						Protocol:                 network.SecurityRuleProtocolTCP,
+						SourceAddressPrefix:      to.StringPtr("*"),
+						SourcePortRange:          to.StringPtr("*"),
+						DestinationAddressPrefix: to.StringPtr("*"),
+						DestinationPortRange:     to.StringPtr("22"),
+						Access:                   network.SecurityRuleAccessAllow,
+						Priority:                 to.Int32Ptr(100),
+					},
+				},
+				{
+					Name: to.StringPtr("inter_node_comm"),
+					SecurityRulePropertiesFormat: &network.SecurityRulePropertiesFormat{
+						Direction:                network.SecurityRuleDirectionInbound,
+						Protocol:                 "*",
+						SourceAddressPrefix:      to.StringPtr("VirtualNetwork"),
+						SourcePortRange:          to.StringPtr("*"),
+						DestinationAddressPrefix: to.StringPtr("VirtualNetwork"),
+						DestinationPortRange:     to.StringPtr("*"),
+						Access:                   network.SecurityRuleAccessAllow,
+						Priority:                 to.Int32Ptr(200),
+					},
+				},
+				{
+					Name: to.StringPtr("azure_load_balancer"),
+					SecurityRulePropertiesFormat: &network.SecurityRulePropertiesFormat{
+						Direction:                network.SecurityRuleDirectionInbound,
+						Protocol:                 "*",
+						SourceAddressPrefix:      to.StringPtr("AzureLoadBalancer"),
+						SourcePortRange:          to.StringPtr("*"),
+						DestinationAddressPrefix: to.StringPtr("*"),
+						DestinationPortRange:     to.StringPtr("*"),
+						Access:                   network.SecurityRuleAccessAllow,
+						Priority:                 to.Int32Ptr(300),
+					},
+				},
+				{
+					// Allow access to node ports from everywhere
+					Name: to.StringPtr("node_ports_ingress"),
+					SecurityRulePropertiesFormat: &network.SecurityRulePropertiesFormat{
+						Direction:                network.SecurityRuleDirectionInbound,
+						Protocol:                 network.SecurityRuleProtocolAsterisk,
+						SourceAddressPrefix:      to.StringPtr("*"),
+						SourcePortRange:          to.StringPtr("*"),
+						DestinationAddressPrefix: to.StringPtr("*"),
+						DestinationPortRange:     to.StringPtr(fmt.Sprintf("%d-%d", portRangeLow, portRangeHigh)),
+						Access:                   network.SecurityRuleAccessAllow,
+						Priority:                 to.Int32Ptr(400),
+					},
+				},
+				// outbound
+				{
+					Name: to.StringPtr("outbound_allow_all"),
+					SecurityRulePropertiesFormat: &network.SecurityRulePropertiesFormat{
+						Direction:                network.SecurityRuleDirectionOutbound,
+						Protocol:                 "*",
+						SourceAddressPrefix:      to.StringPtr("*"),
+						SourcePortRange:          to.StringPtr("*"),
+						DestinationAddressPrefix: to.StringPtr("*"),
+						DestinationPortRange:     to.StringPtr("*"),
+						Access:                   network.SecurityRuleAccessAllow,
+						Priority:                 to.Int32Ptr(100),
+					},
+				},
+			},
+		},
+	}
+
+	updatedRules := append(*parameters.SecurityRules, tcpDenyAllRule(), udpDenyAllRule(), icmpAllowAllRule())
+	parameters.SecurityRules = &updatedRules
+
+	if _, err = sgClient.CreateOrUpdate(a.ctx, cloud.Azure.ResourceGroup, cloud.Azure.SecurityGroup, parameters); err != nil {
+		return fmt.Errorf("failed to create or update resource group %q: %v", cloud.Azure.ResourceGroup, err)
+	}
+
+	return nil
+}
+
+func deleteSecurityGroup(ctx context.Context, cloud kubermaticv1.CloudSpec, credentials Credentials) error {
+	securityGroupsClient, err := getSecurityGroupsClient(cloud, credentials)
+	if err != nil {
+		return err
+	}
+
+	future, err := securityGroupsClient.Delete(ctx, cloud.Azure.ResourceGroup, cloud.Azure.SecurityGroup)
+	if err != nil {
+		return err
+	}
+
+	if err = future.WaitForCompletionRef(ctx, securityGroupsClient.Client); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func tcpDenyAllRule() network.SecurityRule {
+	return network.SecurityRule{
+		Name: to.StringPtr(denyAllTCPSecGroupRuleName),
+		SecurityRulePropertiesFormat: &network.SecurityRulePropertiesFormat{
+			Direction:                network.SecurityRuleDirectionInbound,
+			Protocol:                 "TCP",
+			SourceAddressPrefix:      to.StringPtr("*"),
+			SourcePortRange:          to.StringPtr("*"),
+			DestinationPortRange:     to.StringPtr("*"),
+			DestinationAddressPrefix: to.StringPtr("*"),
+			Access:                   network.SecurityRuleAccessDeny,
+			Priority:                 to.Int32Ptr(800),
+		},
+	}
+}
+
+func udpDenyAllRule() network.SecurityRule {
+	return network.SecurityRule{
+		Name: to.StringPtr(denyAllUDPSecGroupRuleName),
+		SecurityRulePropertiesFormat: &network.SecurityRulePropertiesFormat{
+			Direction:                network.SecurityRuleDirectionInbound,
+			Protocol:                 "UDP",
+			SourceAddressPrefix:      to.StringPtr("*"),
+			SourcePortRange:          to.StringPtr("*"),
+			DestinationPortRange:     to.StringPtr("*"),
+			DestinationAddressPrefix: to.StringPtr("*"),
+			Access:                   network.SecurityRuleAccessDeny,
+			Priority:                 to.Int32Ptr(801),
+		},
+	}
+}
+
+// Alright, so here's the deal. We need to allow ICMP, but on Azure it is not possible
+// to specify ICMP as a protocol in a rule - only TCP or UDP.
+// Therefore we're hacking around it by first blocking all incoming TCP and UDP
+// and if these don't match, we have an "allow all" rule. Dirty, but the only way.
+// See also: https://tinyurl.com/azure-allow-icmp
+func icmpAllowAllRule() network.SecurityRule {
+	return network.SecurityRule{
+		Name: to.StringPtr(allowAllICMPSecGroupRuleName),
+		SecurityRulePropertiesFormat: &network.SecurityRulePropertiesFormat{
+			Direction:                network.SecurityRuleDirectionInbound,
+			Protocol:                 "*",
+			SourceAddressPrefix:      to.StringPtr("*"),
+			SourcePortRange:          to.StringPtr("*"),
+			DestinationAddressPrefix: to.StringPtr("*"),
+			DestinationPortRange:     to.StringPtr("*"),
+			Access:                   network.SecurityRuleAccessAllow,
+			Priority:                 to.Int32Ptr(900),
+		},
+	}
+}
