+++ pkg/provider/cloud/azure/vnet.go
@@ -1,0 +1,83 @@
+/*
+Copyright 2021 The Kubermatic Kubernetes Platform contributors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package azure
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2021-03-01/network"
+	"github.com/Azure/go-autorest/autorest/to"
+
+	kubermaticv1 "k8c.io/kubermatic/v2/pkg/crd/kubermatic/v1"
+)
+
+// ensureVNet will create or update an Azure virtual network in the specified resource group. The call is idempotent.
+func ensureVNet(ctx context.Context, cloud kubermaticv1.CloudSpec, location string, clusterName string, credentials Credentials) error {
+	networksClient, err := getNetworksClient(cloud, credentials)
+	if err != nil {
+		return err
+	}
+
+	parameters := network.VirtualNetwork{
+		Name:     to.StringPtr(cloud.Azure.VNetName),
+		Location: to.StringPtr(location),
+		Tags: map[string]*string{
+			clusterTagKey: to.StringPtr(clusterName),
+		},
+		VirtualNetworkPropertiesFormat: &network.VirtualNetworkPropertiesFormat{
+			AddressSpace: &network.AddressSpace{AddressPrefixes: &[]string{"10.0.0.0/16"}},
+		},
+	}
+
+	var resourceGroup = cloud.Azure.ResourceGroup
+	if cloud.Azure.VNetResourceGroup != "" {
+		resourceGroup = cloud.Azure.VNetResourceGroup
+	}
+	future, err := networksClient.CreateOrUpdate(ctx, resourceGroup, cloud.Azure.VNetName, parameters)
+	if err != nil {
+		return fmt.Errorf("failed to create or update virtual network %q: %v", cloud.Azure.VNetName, err)
+	}
+
+	if err = future.WaitForCompletionRef(ctx, networksClient.Client); err != nil {
+		return fmt.Errorf("failed to create or update virtual network %q: %v", cloud.Azure.VNetName, err)
+	}
+
+	return nil
+}
+
+func deleteVNet(ctx context.Context, cloud kubermaticv1.CloudSpec, credentials Credentials) error {
+	networksClient, err := getNetworksClient(cloud, credentials)
+	if err != nil {
+		return err
+	}
+
+	var resourceGroup = cloud.Azure.ResourceGroup
+	if cloud.Azure.VNetResourceGroup != "" {
+		resourceGroup = cloud.Azure.VNetResourceGroup
+	}
+	deleteVNetFuture, err := networksClient.Delete(ctx, resourceGroup, cloud.Azure.VNetName)
+	if err != nil {
+		return err
+	}
+
+	if err = deleteVNetFuture.WaitForCompletionRef(ctx, networksClient.Client); err != nil {
+		return err
+	}
+
+	return nil
+}
